+++ src/main/java/net/commoble/entitydetectors/Names.java
@@ -1,4 +1,4 @@
-package commoble.entitydetectors;
+package net.commoble.entitydetectors;
 
 public class Names
 {
@@ -6,4 +6,5 @@ public class Names
 	public static final String MOB_DETECTOR = "mob_detector";
 	public static final String IMPRINTED_SLIME_BALL = "imprinted_slime_ball";
 	public static final String FAKE_SLIME = "fake_slime";
+	public static final String IMPRINTED_ENTITY = "imprinted_entity";
 }