+++ src/main/java/net/commoble/entitydetectors/registrables/EntityDetectorBlock.java
@@ -1,10 +1,9 @@
-package commoble.entitydetectors.registrables;
+package net.commoble.entitydetectors.registrables;
 
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.sounds.SoundEvents;
 import net.minecraft.sounds.SoundSource;
-import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.BlockGetter;
@@ -45,38 +44,18 @@ public abstract class EntityDetectorBlock extends Block implements EntityBlock
 		return state.getValue(POWERED) ? 7 : 0;
 	}
 
-	/**
-	 * Can this block provide power. Only wire currently seems to have this change
-	 * based on its state.
-	 * 
-	 * @deprecated call via {@link BlockState#canProvidePower()} whenever possible.
-	 *             Implementing/overriding is fine.
-	 */
-	@Deprecated
 	@Override
 	public boolean isSignalSource(BlockState state)
 	{
 		return true;
 	}
-
-	/**
-	 * @deprecated call via
-	 *             {@link BlockState#getStrongPower(IBlockAccess,BlockPos,EnumFacing)}
-	 *             whenever possible. Implementing/overriding is fine.
-	 */
-	@Deprecated
+	
 	@Override
 	public int getDirectSignal(BlockState blockState, BlockGetter blockAccess, BlockPos pos, Direction side)
 	{
 		return blockState.getSignal(blockAccess, pos, side);
 	}
 
-	/**
-	 * @deprecated call via
-	 *             {@link IBlockState#getWeakPower(IBlockAccess,BlockPos,EnumFacing)}
-	 *             whenever possible. Implementing/overriding is fine.
-	 */
-	@Deprecated
 	@Override
 	public int getSignal(BlockState blockState, BlockGetter blockAccess, BlockPos pos, Direction side)
 	{
@@ -84,8 +63,7 @@ public abstract class EntityDetectorBlock extends Block implements EntityBlock
 	}
 
 	@Override
-	@Deprecated
-	public InteractionResult use(BlockState state, Level level, BlockPos pos, Player player, InteractionHand hand, BlockHitResult hit)
+	public InteractionResult useWithoutItem(BlockState state, Level level, BlockPos pos, Player player, BlockHitResult hit)
 	{
 		if (state.hasProperty(LEVEL))
 		{
@@ -101,7 +79,7 @@ public abstract class EntityDetectorBlock extends Block implements EntityBlock
 		}
 		else
 		{
-			return super.use(state, level, pos, player, hand, hit);
+			return super.useWithoutItem(state, level, pos, player, hit);
 		}
 	}