+++ src/main/java/net/commoble/entitydetectors/registrables/MobDetectorBlock.java
@@ -1,14 +1,13 @@
-package commoble.entitydetectors.registrables;
+package net.commoble.entitydetectors.registrables;
 
-import commoble.entitydetectors.EntityDetectors;
+import net.commoble.entitydetectors.EntityDetectors;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.Containers;
 import net.minecraft.world.InteractionHand;
-import net.minecraft.world.InteractionResult;
+import net.minecraft.world.ItemInteractionResult;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
-import net.minecraft.world.level.block.EntityBlock;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.entity.BlockEntityTicker;
 import net.minecraft.world.level.block.entity.BlockEntityType;
@@ -23,23 +22,20 @@ public class MobDetectorBlock extends EntityDetectorBlock
 	}
 
 	@Override
-	@Deprecated
-	public InteractionResult use(BlockState state, Level world, BlockPos pos, Player player, InteractionHand hand, BlockHitResult hit)
+	public ItemInteractionResult useItemOn(ItemStack stack, BlockState state, Level world, BlockPos pos, Player player, InteractionHand hand, BlockHitResult hit)
 	{
-		ItemStack stack = player.getItemInHand(hand);
 		if (MobDetectorItemHandler.isItemValidFilter(stack) && world.getBlockEntity(pos) instanceof MobDetectorBlockEntity mobDetector)
 		{
 			if (!world.isClientSide)
 			{
 				mobDetector.onRightClickWithSlime(player, stack, world, pos);
 			}
-			return InteractionResult.SUCCESS;
+			return ItemInteractionResult.SUCCESS;
 		}
-		return super.use(state, world, pos, player, hand, hit);
+		return super.useItemOn(stack, state, world, pos, player, hand, hit);
 	}
 
 	@Override
-	@Deprecated
 	public void onRemove(BlockState state, Level world, BlockPos pos, BlockState newState, boolean isMoving)
 	{
 		if (state.getBlock() != newState.getBlock() && world.getBlockEntity(pos) instanceof MobDetectorBlockEntity mobDetector)
@@ -55,6 +51,7 @@ public class MobDetectorBlock extends EntityDetectorBlock
 		return MobDetectorBlockEntity.create(pos, state);
 	}
 
+	@SuppressWarnings("unchecked")
 	@Override
 	public <T extends BlockEntity> BlockEntityTicker<T> getTicker(Level level, BlockState state, BlockEntityType<T> type)
 	{