+++ src/main/java/net/commoble/entitydetectors/registrables/PlayerDetectorBlock.java
@@ -1,9 +1,8 @@
-package commoble.entitydetectors.registrables;
+package net.commoble.entitydetectors.registrables;
 
-import commoble.entitydetectors.EntityDetectors;
+import net.commoble.entitydetectors.EntityDetectors;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.level.Level;
-import net.minecraft.world.level.block.EntityBlock;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.entity.BlockEntityTicker;
 import net.minecraft.world.level.block.entity.BlockEntityType;