+++ changelog.md
@@ -1,3 +1,9 @@
+# 1.21.1-4.0.0.0
+* Updated to Minecraft 1.21.1. Requires neoforge 21.1.73 or higher.
+
+# 1.20.1-3.0.0.0
+* Updated to Minecraft 1.20.1. Requires forge 47.1.0 or higher.
+
 # 1.19.2-2.0.0.0
 * Slimeballs will now only imprint when used on mobs if the player is crouching.
 * Any item in the "entitydetectors:imprintable" item tag can now be used to make mob imprints. By default this includes all items in the "forge:slimeballs" item tag.